The preproduction version. Just still some handles for images upload and download and image holder for course and instructor. And want to replace some normal get data functions with firebase listeners.

+++ app/src/main/java/com/kareimt/anwarresala/presentation/viewmodels/CoursesViewModel.kt
@@ -4,13 +4,14 @@ import android.content.Context
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
+import androidx.compose.ui.text.intl.Locale
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
 import com.kareimt.anwarresala.data.local.branch.BranchEntity
 import com.kareimt.anwarresala.data.local.course.CourseEntity
 import com.kareimt.anwarresala.data.local.DatabaseProvider
-import com.kareimt.anwarresala.data.remote.repository.course.CourseRepositoryInterface
+import com.kareimt.anwarresala.domain.repository.course.CourseRepositoryInterface
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -55,11 +56,11 @@ class CoursesViewModel(context:Context, private val repository: CourseRepository
         viewModelScope.launch(Dispatchers.IO) {
             setLoading(true)
             try {
-                val query =_uiState.value.searchQuery.trim()
+                val query =_uiState.value.searchQuery.trim().lowercase(java.util.Locale.getDefault())
                 val searchResults = if (query.isEmpty()){
                     courseDao.getAllCourses()
                 }else{
-                    courseDao.searchCourses("%$query%")
+                    courseDao.searchCourses(query)
                 }
                 updateCourses(searchResults)
             } catch (e: Exception) {
@@ -70,7 +71,7 @@ class CoursesViewModel(context:Context, private val repository: CourseRepository
         }
     }
 
-    // TODO: Complete the Cycle of adding and removing branches
+    // Handling Branches
     private val _branches = MutableStateFlow<List<BranchEntity>>(emptyList())
     val branches: StateFlow<List<BranchEntity>> = _branches.asStateFlow()
 
@@ -103,6 +104,7 @@ class CoursesViewModel(context:Context, private val repository: CourseRepository
         }
     }
 
+    // Handling Courses
     private val _courses = MutableStateFlow<List<CourseEntity>>(emptyList())
     val courses: StateFlow<List<CourseEntity>> = _courses
 
@@ -112,7 +114,9 @@ class CoursesViewModel(context:Context, private val repository: CourseRepository
 
     init {
         loadCourses()
+        setCoursesListener()
         loadBranches()
+        setupBranchesListener()
     }
 
     private fun loadCourses() {
@@ -122,44 +126,91 @@ class CoursesViewModel(context:Context, private val repository: CourseRepository
     }
 
     fun addCourse(course: CourseEntity) {
+        setLoading(true)
         viewModelScope.launch(Dispatchers.IO) {
-            courseDao.insertCourse(course)
-            // Update both flows
-            _courses.value = courseDao.getAllCourses()
-            updateCourses(courseDao.getAllCourses())
+            try {
+                repository.addCourseOnFirebase(course).fold(
+                        onSuccess = {
+                            courseDao.insertCourse(course)
+                            loadCourses()
+                            setLoading(false)
+                            updateMessage("Course added succesfully")
+                        },
+                        onFailure = { error ->
+                            setError("Failed to add course: ${error.message}")
+                            setLoading(false)
+                        }
+                    )
+            } catch (e: Exception) {
+                setError("Exception: ${e.message}")
+                setLoading(false)
+            }
         }
     }
 
     // For update exact course
     fun updateCourse(course: CourseEntity) {
+        setLoading(true)
         viewModelScope.launch(Dispatchers.IO) {
-            courseDao.updateCourse(course)
-            loadCourses()
+            try {
+                repository.updateCourseOnFirebase(course).fold(
+                    onSuccess = {
+                        courseDao.updateCourse(course)
+                        loadCourses()
+                        setLoading(false)
+                        updateMessage("Course updated successfully")
+                    },
+                    onFailure = { error ->
+                        setError("Failed to update course: ${error.message}")
+                        setLoading(false)
+                    }
+                )
+            } catch (e: Exception) {
+                setError("Exception: ${e.message}")
+                setLoading(false)
+            }
         }
     }
 
     // For Delete exact course
     fun deleteCourse(course: CourseEntity) {
+        setLoading(true)
         viewModelScope.launch(Dispatchers.IO) {
-            courseDao.deleteCourse(course)
-            loadCourses()
+            try {
+                val result = repository.deleteCourse(course.id)
+                if (result.isSuccess) {
+                    courseDao.deleteCourse(course)
+                    loadCourses()
+                    withContext(Dispatchers.Main) {
+                        setLoading(false)
+                        updateMessage("Course deleted successfully")
+                    }
+                } else {
+                    withContext(Dispatchers.Main) {
+                        setLoading(false)
+                        setError("Failed to delete course: ${result.exceptionOrNull()?.message}")
+                    }
+                }
+            } catch (e: Exception) {
+                withContext ( Dispatchers.Main) {
+                    setLoading(false)
+                    setError("Exception: ${e.message}")
+                }
+            }
         }
     }
 
     // For Getting course by ID
     private val _selectedCourse = MutableStateFlow<CourseEntity?>(null)
     val selectedCourse: StateFlow<CourseEntity?> = _selectedCourse.asStateFlow()
-    fun getCourseById(courseId: Int) {
+    fun getCourseById(courseId: String) {
         viewModelScope.launch(Dispatchers.IO) {
             val course = courseDao.getCourseById(courseId)
             _selectedCourse.value = course
         }
     }
 
-
-
-
-    // After add firebase
+    // New only After add firebase
     // Attributes
     var onError by mutableStateOf("")
         private set
@@ -174,7 +225,7 @@ class CoursesViewModel(context:Context, private val repository: CourseRepository
 
 
     // Normal functions
-
+    // Branches functions
     fun addBranch(newBranch: String) {
         setLoading(true)
         try {
@@ -232,6 +283,29 @@ class CoursesViewModel(context:Context, private val repository: CourseRepository
                 }
             }
         }
+    }
+
+    fun setCoursesListener(){
+        repository.getCourses { result ->
+            result.fold(
+                onSuccess = { courses ->
+                    viewModelScope.launch (Dispatchers.IO){
+                        courseDao.deleteAllCourses()
+                        courseDao.insertCourses(courses)
+                        _courses.value = courses
+                        println("Successfully loaded ${courses.size} courses from the internet")
+                    }
+                },
+                onFailure = { error ->
+                    println("CourseSync, Error: ${result.exceptionOrNull()}")
+                    _uiState.update { it.copy(error = error.message) }
+                }
+            )
+        }
+    }
+
+    fun removeCoursesListener() {
+        repository.removeCoursesListener()
     }
 }
